+++ indico/web/client/js/jquery/widgets/jinja/index.js
@@ -31,3 +31,4 @@ import './typeahead_widget';
 import './permissions_widget';
 import './sui_color_picker_widget';
 import './sui_remote_search_dropdown_widget';
+import './tinymce_widget';